Add integration test for filteredfs migration
Summary:
Create a testing suite for FilteredFS migration.

Added a bunch of utilities to this class so we can use them in future diffs when we actually implement and enable the migration. So most of the functions added in this diff are not doing interesting things. Only a simple test case is added for now. More will be added in future diffs.

Reviewed By: MichaelCuevas

Differential Revision: D85066125

fbshipit-source-id: 1bf17f9d43ff47f2e8c4220f7b91a1950bc1630b