Add description tooltips
Summary:
I think it could be helpful to show some kind of description when you hover over a smart action, both in the primary button and in the menu. I added some handling for both the `ButtonDropdown` and `ContextMenu` components to allow for this.

For the context menu, I think it makes sense to only allow a tooltip for items with no sub-items.

NOTE: Descriptions can probably be tweaked.

Reviewed By: evangrayk

Differential Revision: D85355889

fbshipit-source-id: 778fc04fb93f8e0d2fbe4ad210bbe2bc77f2e279