Fix offset for context menu spawn
Summary:
I noticed a tiny bug where the context menu appears slighly offset from the mouse cursor position in the bottom-right quadrant of the webview. I think this is because of an incorrect operator when attempting to calculate the offset.

I noticed this when working on the smart actions component, which also uses the context menu. I added specific offsets in this diff for that component too, so that the context menu appears directly above/below the dropdown button.

Reviewed By: evangrayk

Differential Revision: D85286868

fbshipit-source-id: b7616b6f719c6f9aa2ef432a2e05bcfeecf04964