Re-order actions
Summary:
This diff re-arranges the initial order of the smart actions to prefer certain actions over others. Specifically, we want to show diff actions first, if available, and prefer actions that we want to boost discoverability for (e.g. code review).

We don't need to move the merge conflict action at all since it's a special case and will be the only action to appear in the merge conflict state.

Reviewed By: evangrayk

Differential Revision: D85286866

fbshipit-source-id: dce6472c7796c2a08fd7ca517ed5575198268de0