Add custom select menu component
Summary:
`ButtonDropdown` uses a basic `select` element to render available options, but I wanted the dropdown for the smart actions component to look more "native" to ISL while still being streamlined compared to the menu of buttons for the old smart actions component.

A good candidate here, at least visually, is actually the context menu component that appears when right-clicking in the ISL main window. It has enough stuff built-in for us to extend the functionality of the smart actions menu later (dividers, sub-menus, etc.).

I needed to add a new property to the `ButtonDropdown` component to support a "custom" select component, which is a bit janky but seems to work well enough. I also am using the bounding box for the button instead of the cursor position to determine where to spawn the menu, which is appropriate here.

Reviewed By: evangrayk

Differential Revision: D83487380

fbshipit-source-id: 673c69a93d6f72b4da206e8251f96c1e535cb960