Improvements to perf / feature flag fetching
Summary:
Earlier in this stack I introduced some dynamic feature flag fetching, which violates the rules of hooks. I sort of waved it away with a claim that it shouldn't cause issues since the list of smart actions is static and constant (meaning they'll load in the same order every time) -- this is sort of true, but having this unfortunately breaks hot-reloading for the whole app.

To fix this, I'm making a few significant changes in this diff:
- Using `bulkFetchFeatureFlags` to avoid conditionally fetching hooks in a loop, via a separate atom
- Wrapping that atom in a loadable to properly suspend
- Adding a `useEffect` hook to set the selectedAction after actions have finished loading
- Adding `useMemo` hooks to memoize the action context and available actions
- Moving all hooks ahead of conditional rendering clauses

This should avoid the hooks violation and re-enable hot reloading.

Reviewed By: evangrayk

Differential Revision: D85286865

fbshipit-source-id: 23fb9dbd6d0884fc1f4bf5d347d6f93e01468c72