Add smart action for resolving conflicts
Summary:
To consolidate AI offerings in the ISL, specifically in the commit tree, it makes sense to combine the "resolve conflicts" action into the existing list of actions. I made a few changes to accomodate this while preserving the behavior of the button that stbfb added.
- Added a new "merge-conflicts" smart action config type as a special case. If there are merge conflicts, we only show the merge conflict actions. Otherwise, we never show these actions.
- If only one action is shown, we use a normal button instead of the `ButtonDropdown`.

The main differences now are that I'm using a subtler "icon" button (for consistency with the dropdown, but we could change this) and the lightbulb icon (which we discussed could also change).

Reviewed By: stbfb

Differential Revision: D85286867

fbshipit-source-id: e254147b1d8f7159d010e5885bc03496b3320f78