Ensure smart actions are only shown internally
Summary:
Currently, only GKs are used to ensure that smart actions are only shown internally, but we'll want to eventually remove these once fully rolled-out.

We can add an additional check to ensure we only show smart actions internally. (In the future, we may want to change this).

In a future diff I'll look towards moving the entire smartActions/ directory to internal code as well.

Reviewed By: evangrayk

Differential Revision: D85286864

fbshipit-source-id: 839a92aebec1d83c6f73845bc1767ae6844d9f23