Add actual actions to dropdown
Summary:
Here I'm defining a constant array that will serve as the source-of-truth for smart actions going forward, as well as types to support this. There are some common traits that we can define for each smart action:
- tracking event
- GK
- platform restriction (really only relevant to 'vscode' right now since most if not all actions require DVSC)

I also try to standardize when we show certain smart actions (using the above traits, plus custom logic if any) and how they execute. There are a few cases where I need to do some initial work before running the action, which I leave up to `getMessagePayload`.

There are also 3 different types of smart actions:
- `simple`: appear everywhere (commit, diff, uncommitted changes)
- `commit`: only appear on commits (and diffs)
- `diff`: only appear on diffs

I thought there might be more substantial differences between the various types, but I end up only using this for a check in `shouldShowSmartAction`. There's also some custom logic for one smart action where we only show it on uncommitted changes. I figured this leaves room to add things later, but I'm not attached to the way that it's defined right now.

For each of the smart actions, I try to preserve the behavior defined in the original component. Overall, this system seems to work fairly well for the current set of smart actions, but we may need to make adjustments later to allow for added flexibility.

Reviewed By: evangrayk

Differential Revision: D85286863

fbshipit-source-id: bf320e0904a3d7816c0de26a1224449e238bd639