transition to AuthenticatedIdentities
Summary:
MononokeIdentity pretty much just mirrors the standard Identity structure - both hold two strings: id_type and id_data. We want to track extra data, for example whether an action is being made by a real user, or whether it is an AI agent on behalf of the user. The AuthenticatedIdentity provides that. It's also a superset of the data in Identity - both contain id_type and id_data, but the AuthenticatedIdentity provides extra attributes. Let's migrate to it.

The plan is to in the end have MononokeIdentity just contain the AuthenticatedIdentity structure, which has all the data inside of it as codegened **Rust** types (not pointers to some cpp structs). We will keep accessors like `id_type()` and `id_data()` compatible.

At first I wanted for time of the transition to dry-run the change at runtime, and check whether the identities in the old way of parsing are the same as in the new way. I wanted to log when they're weren't.

However as per [this commend](https://www.internalfb.com/diff/D85012684?dst_version_fbid=25449375208019436&transaction_fbid=848915844254781), we can only expect one or the other types of identities - not both. This means we can't do that. We can however fall back to the usual way if the authenticated identities aren't there.

We need a switch between the two for a gradual, safe rollout, and potentially quick rollback - this is what the JK is for.

Once for a long-enough period of time we see everything is compatible, we'll just clean all of this up, and have something like:
```
struct MononokeIdentity {
   auth_ident: AuthenticatedIdentity,
}
```

(or perhaps make MononokeIdentity a trait. for now we need both to do runtime checks)

Reviewed By: clara-9

Differential Revision: D85012684

fbshipit-source-id: 16ab3abd7a38d92fe7317dc47fc15395de416c89