Make rebase warning check synchronous
Summary:
With the removal of the rebase warnings GK in 0.1.70009, [there are no more rebase timeouts](https://fburl.com/scuba/sapling_isl_analytics/osu97lvc).

Let's make the check to see if a warning should be shown synchronous, and get rid of the timeout. The previous code was also a bit unnecessarily verbose, and can be cleaned up a good amount (next diff).

Reviewed By: evangrayk

Differential Revision: D85533527

fbshipit-source-id: 1278fd8c824514fc2367d161295bfb2836600ae4