Show dropdown button persistently on uncommitted changes
Summary:
Somehow I missed this in my initial stack, but for uncommitted changes the smart actions dropdown renders in the same place as the old smart actions menu, so it's not persistent. If I try to make it persistent in that button row, it either would need to be before the quick commit field (which I don't think makes sense) or appear to the right of non-persistent buttons (which should be an obvious no-go).

Real-estate is limited here, so I decided to put it in the top row (with discard, absorb, etc.) instead. Not in love with it but functionally I think it aligns better here.

Reviewed By: evangrayk

Differential Revision: D85594675

fbshipit-source-id: e4442c109f0f40b9d03f4b2b80009d075cf4c80b