Fix race when fetching files
Summary:
There was a race condition when fetching files. The commit info sidebar has to fetch the full list of changed files when you open it, as the CommitInfo only contains a small subset of files.

This fetch is async and takes time, and is keyed by hash.

If you kick off a fetch on commit A, then switch to a new commit B, it would kick off another fetch for that commit. If that one finishes very fast, it may return commit B's results before commit A. But the old implementation didn't cancel the fetch, so when commit A's results came it, it would set the state and overwrite commit B's files.

The fix is simply to use a cancellation token to ensure we don't set state if the hash has changed.

Reviewed By: babyleafy

Differential Revision: D85595867

fbshipit-source-id: b1bc5ddccd3adca91629871910eacc6445320e88