git/protocol: rewrite packfile_stream_from_changesets method to spawn two tasks instead of use custom stream
Summary:
Rewrite the `packfile_stream_from_changesets` method to not spawn two tasks instead of creating a custom stream. The main challenge with this method is that it processes each changeset fetching multiple GDM entries for each, and for each GDM entry fetches/creates a packfile item. The first mapping being one-to-many complicates using standard stream buffering which is why we use two `FuturesOrdered` and try to balance them explicility.

Although this approach is faster than chaining `.buffered`, it still does not yield optimal performance. This diff switches to spawning multiple tasks and using channels to send the GDM entries/packfile items, which is both simplier and more efficient.

Reviewed By: RajivTS

Differential Revision: D85674217

fbshipit-source-id: 4b75c55b48f1b7170ca01d178e986bd7c9c71901