don't add to the counter on success
Summary:
We don't depend on this counter for alerts, so it's fine for it to go missing.

Becuase we bump it with value 0 this makes it confusing for avg aggfregation as it shows 0.

With this change the counter might not return data if there were no failures, but samples wil indicate failures,w hich makes it easier to use in canaries

Reviewed By: RajivTS

Differential Revision: D85677391

fbshipit-source-id: 039d730e6b36144fdc52bb91114d97f88c6d0cb1