snapshots: optimize the creation flow to avoid per commit lookup
Summary:
optimize the creation flow for snapshots

Right now on each creation we validate the backed up property for the entire
stack on remote. However, we could entirely avoid it by using the local
backedup cache in the same way as commit cloud uses.

The win should be around 100ms as we will be skipping the lookups that are on the backend translated to XDB/memcache checks.

Reviewed By: muirdm

Differential Revision: D85563108

fbshipit-source-id: 50cae99a4f771f8846a2242730be5795094d3ca1