backingstore: pass matcher lambda to prefetch_manager
Summary:
# This diff

Instead of passing in the matcher directly to the prefetch manager, we should instead pass a lambda which can be used to "refresh" the sparse matcher that's used to determine the sparseness of files/dirs in the repo. This allows the sparse matcher to evolve over time as the user checks out different commits.

Previously, the same matcher would be used regardless of whether a new commit was checked out, which was incorrect in the cases that the sparse profile was updated between commits.

# Context

See D85361094

Reviewed By: muirdm

Differential Revision: D85382468

fbshipit-source-id: d1a203e8d7b8d0e0f8f460cb1ca1c67641cc0936