New API for diffing while excluding generated files
Summary:
Add a new API for diffing. The previous API was limited to diffing against a specific commit. Instead, we should expose the concept of `Comparison`s, which is how the rest of ISL does diffs and such. This is more powerful as it can be used in different ways.

Additionally, add an `excludeGenerated` option to filter the diff patch to remove generated files. This parses the output, then uses our generated file detection to find which paths are generated, and removes them.

Since we're using generated files, we may as well also expose a function to learn about which files are generated in the API. This is just a simple filter function.

Differential Revision: D85287904

fbshipit-source-id: 9f031a049f3442ac64e51cd4a0eebe03b7476be8