Stringify and filter diff patches
Summary:
In this stack, we will filter diff patches to exclude generated files (for use in the API).

To do this, we must be able to parse and then stringify a diff patch, so we can filter in between. This diff adds a stringify function, and a helper to filter that uses it.

Reviewed By: taidaii

Differential Revision: D85287901

fbshipit-source-id: 05996e6ada0e93868934d75aac1380a5688a9209