snapshots: add an option to reuse existing TTL of the bubble for Snapshot Continuation
Summary:
add an option to reuse existing TTL of the bubble for Snapshot Continuation feature

The time spent on extend bubble ttl is minimal (~140 ms per invocation) and
includes XDB write. By skipping it, we would store the snapshots faster for the devmate usecase, while the TTL for the chain will be equal to the TTL of the snapshot that initially formed the chain.

Please see D85563108 for perf breakdown

**Usage scenario:**

*Initial Snapshot Creation:*
Devmate executes the first snapshot creation command, specifying the desired lifetime. This is done as the chain of snapshots begins to form.

*Subsequent Snapshots:*
For all following snapshot creations:
The command includes the `--continuation-of <snapshot id>` flag to link to the previous snapshot.

The `--keep-ttl` flag is also added, which bypasses the TTL extension logic. This is because the snapshots are created in quick succession, making TTL extension unnecessary.

Reviewed By: muirdm

Differential Revision: D85675209

fbshipit-source-id: 77466140bca87dbda90699090d9e49737ad284bf