Change TreeEntry to keep executable bit on Windows
Summary:
# This diff:
This is the actual diff that implements the changes. The `getType` should return the `type_` of a TreeEntry, and it should not mask EXECUTABLE file types. Also, the `treeEntryTypeFromMode` should return EXECUTABLE type if the EXECUTABLE permission bit is set even on Windows.
### Note 1:
`S_IXUSR` is zero on Windows because executable doesn't have any meaning on Windows. However, I defined `TREE_ENTRY_EXECUTE_BIT` to use as a replacement for `S_IXUSR`, and we can set the same value for Windows and non-Windows platforms. In this way, inside eden, we consider file mode can be EXECUTABLE, even on Windows. Therefore, we can a consistent file's mode and executable permissions on all platforms.
### Note 2:
The changes on `getType()` are under closed config. Then it doesn't break anything, we will gradually open this config. Also, the changes on `treeEntryTypeFromMode()` is not important when the config is closed. Because when the config is close eden mask the executable permission bit on the file's mode, therefore the `treeEntryTypeFromMode()` won't return EXECUTABLE on windows while the config is close.

Reviewed By: muirdm

Differential Revision: D84554581

fbshipit-source-id: 81295c92271904a1a8ab71c1ea1f3e107559312c