Integration test to check hg status for Executable files on Windows
Summary:
This is an integration test to check the behavior of `hg status` on Windows after we add remembered executable bit ability to Windows.
Consider this test case:
- a file is executable on source control
- Windows fetch the file. Due to the current changes in this stack edenFS on Windows will know the file is executable same as other platforms.
- We delete the file localy and create the file again. But this time the file is regular (It is not possible to create an executable file on Windows).

This is an expected behavior and the test could pass on all platforms.

To create a test repo with a executable file on windows it uses `debugmakeexecutable` command to mark the file in the repo manifest as executable

Reviewed By: muirdm

Differential Revision: D85475526

fbshipit-source-id: 74bcf1a8149dd51cf82e75decb863fc2529fc5e3