add compareTreeEntryType
Summary:
# This diff:
There are some special cases that `sl st` can return some unchanged files when the `windowsRememberExecutableBit` config is true.
For example, when user changes an executable file and then revert the changes with `sl revert --all` the file will written on disk as a new file, and because Windows doesn't have Executable permission bit the new file written on disk is REGULAR_FILE while the same file on source control is EXECUTABLE file. This discrepancy could result to mode changes that can below up in `sl st`

To fix these special cases:
- This diff creates a new function in `TreeEntry` called `ignoreWindowsExecutableTypeForComparison` this method filter file's type, return REGULAR type for EXECUTABLE type only on Windows.
- On the places that we compare file types, we call this function to filter out EXECUTABLE type on Windows.
- On the places that we want to set the initial mode or we want the actual source control file type for thrift request, we don't use this method and the previous methods works as expected.

Reviewed By: muirdm

Differential Revision: D85198139

fbshipit-source-id: 921b17171ba6465842f433a703a1385383bb4617