add windowsRememberExecutableBit to some places
Summary: Next diff needs the `windowsRememberExecutableBit` config in a hot path (it will run for every tree load). Instead of loading config on that place and finding the value of `windowsRememberExecutableBit` config which could be time consuming, we can add a new private variable to these classes and use when it is needed.

Reviewed By: muirdm

Differential Revision: D84945367

fbshipit-source-id: 8ecebec0c94bc0a1a152f3832bddf1e1f334d6cb