Add option to print position as a string instead of a struct
Summary:
Adds new flag --formatted-position to eden notify changes-since

Why:
Non-rust clients based on the CLI interface over json. Specifically for the node client, mount_generation is too big to fit into a normal int and gets truncated. Since we intend for position to be an opaque value for end clients, add an option for returning position back as a string.

Reviewed By: zzl0

Differential Revision: D85308901

fbshipit-source-id: 64914902b9d11c3a534818a9cdf298012d55bee0