Remove Circular Dependencies (Add mergeConflicts/consts.tsx)
Summary:
Working on D84986618, trying to import `FileTree.tsx` throws a lot of errors related to circular imports/dependencies. (Mostly in the format of "Cannot access `Internal.ts` before initialization...")

Continuing with this diff stack, there is another circular dependency created through this loop:

```
mergeConflicts/state.tsx
-> Internal.ts
-> InternalImports.ts
-> FullRepoBranchMergeUtils.tsx
-> FullRepoBranchMergeModal.tsx
-> FileTree.tsx
-> ChangedFile.tsx
-> mergeConflicts/state.tsx
```

As these only seem to import `CONFLICT_SIDE_LABELS`, we move this out to its own file (`consts.tsx`). I'm open to renaming this file or moving it in a different file or directory.

## Future

Any component importing `Internal.ts` is at risk of introducing a circular dependency. For example, there is another circular dependency in `BookmarksData.tsx`:

```
BookmarksData.tsx
-> Internal.ts (importing feature flags and Internal.getRecommendedBookmarks)
-> InternalImports.ts
-> BookmarksData.tsx (importing recommendedBookmarksAtom)
```

But this is not trivial to resolve as both files are tightly coupled and I don't have insight as to what these methods are doing.

I'm not sure what the long-term solution would be for circular dependencies introduced through `Internal.ts` — maybe making sure all components imported by Internal are also Internal (and if not, making a copy of it to Internal).

Reviewed By: evangrayk

Differential Revision: D85538058

fbshipit-source-id: 0113b5e1e33e7952f26296eaba863c221dbbe460