Add BookmarkState to allow tracking of more than one CS pointer for a bookmark
Summary:
### Background
The Bookmarks cache can be configured with various sets of warmers, depending on the specific data requirements of each service.

Currently, different services use different warmer sets, but the Bookmarks service only exposes the Hg set. As a result, these services cannot stop using the Bookmarks cache directly.

With this stack, we are extending the Bookmarks cache to track three distinct pointers: Hg, Git, and all configured warmers. The goal is for the Bookmarks service to configure all warmers and expose all three pointers through its API, enabling different clients to access the cache directly according to their needs.

### This change
In this change we create a  BookmarkState struct to track the different sets per scope. At this stage there is no change in behaviour, we only actually track a single changeset (for all warmers), so that the cache can keep on working as before. The remaining CS pointers will come in the next diffs.

Reviewed By: gustavoavena

Differential Revision: D85769790

fbshipit-source-id: 2f53ba4da130aacff75ea42f4aed5f7bf287b800