Introduce scoped trait
Summary:
### Background
The Bookmarks cache can be configured with various sets of warmers, depending on the specific data requirements of each service.

Currently, different services use different warmer sets, but the Bookmarks service only exposes the Hg set. As a result, these services cannot stop using the Bookmarks cache directly.

With this stack, we are extending the Bookmarks cache to track three distinct pointers: Hg, Git, and all configured warmers. The goal is for the Bookmarks service to configure all warmers and expose all three pointers through its API, enabling different clients to access the cache directly according to their needs.

### This change
In this change we will create a new trait that allows clients to request the particular scope they care about. In coming diffs we will start to track these pointers, so that we can return the one that the user needs.

Note that the previous unscoped trait will stay, and will keep on working exactly the same as today.

Reviewed By: gustavoavena

Differential Revision: D85769783

fbshipit-source-id: 8bff7389d689e9fbb5e0d8b68c9f0abaa7ac632b