Add warmer tags, including microwave
Summary:
### Background
The Bookmarks cache can be configured with various sets of warmers, depending on the specific data requirements of each service.

Currently, different services use different warmer sets, but the Bookmarks service only exposes the Hg set. As a result, these services cannot stop using the Bookmarks cache directly.

With this stack, we are extending the Bookmarks cache to track three distinct pointers: Hg, Git, and all configured warmers. The goal is for the Bookmarks service to configure all warmers and expose all three pointers through its API, enabling different clients to access the cache directly according to their needs.

### This change
Since users will need to configure different sets of warmers for different purposes (Git / Hg), we need to tag them (they can belong to either set, or to both of them). At this stage the tag is not used for anything, just included in the Warmer.

Reviewed By: gustavoavena

Differential Revision: D85769787

fbshipit-source-id: de9996e7134a774f8b7c6f38c0f928b9305f1486