TreeInode: extend lazyInodePersistence to checkout
Summary:
Apply the lazyInodePersistence flag during checkout to skip writing directories to the overlay when they aren't materialized.

There seem to be two reasons we write non-materialized directories to the overlay:
1. As explained in D7598094, we store inode metadata (e.g. mode) by inode number, so remembering inode numbers is important to keep track of the metadata. However, before writing the metadata we materialize the file and all ancestor directories, which should guarantee inode persistence (i.e. it doesn't seem like we need to write non-materialized directories to the overlay for the purpose of future metadata updates).
2. The other reason is that some programs (such as clang) are known to temporarily store inode numbers to perform file identity checks. Even if the file isn't open and all kernel caches are empty (i.e. inode is completely unreferenced), eden is expected to remember the file's inode number for a period.

lazyInodePersistence is a flag that causes eden to defer non-materialized directory writes to unload time instead of eagerly writing them to the overlay at load time. This keeps the desired behavior for 2) while avoiding the cost of writing directories to the overlay in the read path. This commit applies the same optimization to the checkout path, which similarly was eagerly writing directories to the overlay (even if not materialized).

The cost of the optimization is that:
1. Shutdown (i.e. graceful restart) is slow because we have to write out all the directories to the overlay (laziness catches up with us).
2. We will forget non-materialized inodes if we crash.

Reviewed By: genevievehelsel

Differential Revision: D85914644

fbshipit-source-id: 3803c4d491b64eede979bb7174b6e1974e65f1db