Don't use ==== to detect conflicts
Summary:
If you have `=======` in your file and you're resolving conflicts, it would see that and think you still have unresolved conflicts. Let's just lean on the `<<<<<<<` and `>>>>>>>` and `|||||||` to detect this, since `====` can be used a separator in comments and markdown and stuff.

VS Code's conflict detection highlighting only triggers if you have ALL 3 of these markers present.

This could cause conflicts to get auto-resolved when you didn't intend if you removed parts of the conflicts but still had the ==== and saved. But since you still have to manually click `sl continue` I don't think this is dangerous.

I suppose we could instead just make sure to only count exactly 7 `=`, so that you'd have to get really unlucky with your `=====` to be the right size to reduce false positives....

Reviewed By: zzl0

Differential Revision: D86131287

fbshipit-source-id: 5e4fa27bfc412bddcb5aafd6803966e64a9e5b53