edensparse migration: Do not create the marker file if migration is not needed
Summary:
The migration to edensparse should be idempotent so we don't have any side effect when the repo is already filtered. This is done by making sure nothing from the step 2 of the migration cause any changes.
This diff modifies the return type of these functions:
```
- create_empty_sparse_file(checkout)
- update_hg_requires(checkout)
- configure_sapling(checkout)
```
so they return a boolean indicating if they actually make changes. If any of them makes a change, that means the checkout is not filteredfs-ready (otherwise no changes needed).

All these above is needed to avoid creating the marker file when the repo is filteredfs-ready.

Reviewed By: MichaelCuevas

Differential Revision: D85621315

fbshipit-source-id: ab64f650563143cf359344c20536d469e1cd26ae