edensparse migration: Duplicate test cases with a new commit
Summary:
Before this diff, all testing in this test suit is testing with the initial commit.
That means the SNAPSHOT file is always with format version 2.
This diff duplicates the test cases by testing after making a second commit,
forcing the SNAPSHOT file in format version 4.

So we can test the SNAPSHOT file update logic of `update_snapshot_file` in utils.py.

Reviewed By: MichaelCuevas

Differential Revision: D85699356

fbshipit-source-id: 98edfc19e3c5d9b88b5d23129f776ccdea49bf82