edensparse migration: Add rollbacks
Summary: Migration operations can fail at any point. Instead of leaving the repo with a modified state, it's better to restore the repo back to the original state before the migration. To achieve this, we register a rollback function every time we make a change. And if any exception is captured, we rollback the changes.

Reviewed By: MichaelCuevas

Differential Revision: D85772166

fbshipit-source-id: 7aef50f3d967d3b39fbea8e49b526a1090c0a802