code_tenting: Disable manifest id cache if the repo doesn't have restricted paths
Summary:
We're making a lot of DB queries, so I looked into it and confirmed that it's because we're making queries for all the repos, regardless of if they have any restricted paths: https://fburl.com/scuba/mononoke_sql_telemetry/iobl038x

This is likely contributing the [the CPU spikes](https://fburl.com/canvas/uvqkaw8f) that are leading to queries being killed by pylander.

This should address the issue, but only creating the cache if the repo has restricted paths.

Reviewed By: lmvasquezg

Differential Revision: D86433019

fbshipit-source-id: 54df392e9d476114848a995a253f369387ede1ad