Remove --no-prefix arg on blobstore subarg
Summary:
Now raw_blobstore subcommand supports all operations that blobstore subccommand does.

In order to make the transition to KeyedBlbostore better, we need to remove the possiblity of creating a KeyedBlobstore with no repo_id, so I'm removing the possibility of using storage config or `--no-prefix`. Instead, that functionality is supported via the `raw-blobstore` subcommand

Differential Revision: D86312251

fbshipit-source-id: e857cedab053082950d6627fc1bd6790e2f6d15a