pyedenapi: add tracing spans around non-streaming endpoints
Summary:
This way random endpoints like the commit cloud ones will show up in perftrace when they are slow.

I added the tracing in the pyedenapi layer because the it is easier than the underlying Rust async functions. My main goal is to make sure slow SLAPI requests show up in perftrace output during Python commands.

Adding tracing for streaming results is a bit more nuanced, so skip for now.

Reviewed By: zzl0

Differential Revision: D86370299

fbshipit-source-id: 91ffd079030561bcac16cf0d7f7a92baafff36ae