chenge the default value of chuncking takeover data to true
Summary:
This config has been open to all for more than 5 months. I want to remove this config to make the graceful restart code easier for readers.
https://www.internalfb.com/code/configerator/[21e07f871a54]/source/scm/eden/config/edenfs_config.cconf?lines=66-68

To be more caution and confirm that it doesn't break anything. I changed the default value of this config to true. It will be part of next week release and if everything goes well, we can safely remove this config.

Reviewed By: genevievehelsel

Differential Revision: D86546274

fbshipit-source-id: f1e71369a191894833b90fde9702741de376ee8c