Make repo_names_in_tier include hash scheme
Summary:
I learnt that this
https://www.internalfb.com/code/fbsource/[4d8795a7de64a3063f2e2877dcfd40f097c8f101]/fbcode/eden/mononoke/scs/scs_methods/src/methods.rs?lines=42-44
only contains repos that are served by that specific shard, not all repos, so I can't use that as the base for the filter.

In this diff, I'm extending the `repo_names_in_tier` to include identity scheme as well, so the `list_repos` method could filter there.

Reviewed By: markbt

Differential Revision: D86596572

fbshipit-source-id: c48277dab80ff67d73e9be56c4eea55337cca8ec