sql_ext: Start logging with new logger
Summary:
# This stack
Defining a logger config to replace the raw logging scuba we're doing for the XDB telemetry.

My plan is to roll this out, log both ways for a week or two to ensure we don't lose data and then keep only the logger once we know the data looks good.

The main blocker to roll this out now is to be able to capture the logs in unit tests, which would let me use the new logger in the existing tests.

## Why?
Schematized loggers are the best way to log. They provide a lot of features and are more reliable than logging to custom scuba categories.

We'll be able to set custom ODS counters, specify custom sampling rates for scuba and hive, define derived columns, etc.

## This diff
With the new logger generated, start logging with it in the telemetry module.

My plan is to log to both Scuba and the schematized logger for a while, because:

1. We can compare both and iterate on the schematized logger until we're confident it's on par with Scuba.
2. We have a long-ish transition period where we have data in both sources. If we do a flag day change, we might want to run queries on the period and we won't be able, because before a certain day we'll only have Scuba and after only logger.

Differential Revision: D86533410

fbshipit-source-id: 03c161419ebee9d198e28647068dd7092e143481