sql_ext: Pass slice of repo ids instead of Vec
Summary:
# This stack
Defining a logger config to replace the raw logging scuba we're doing for the XDB telemetry.

My plan is to roll this out, log both ways for a week or two to ensure we don't lose data and then keep only the logger once we know the data looks good.

The main blocker to roll this out now is to be able to capture the logs in unit tests, which would let me use the new logger in the existing tests.

## Why?
Schematized loggers are the best way to log. They provide a lot of features and are more reliable than logging to custom scuba categories.

We'll be able to set custom ODS counters, specify custom sampling rates for scuba and hive, define derived columns, etc.

## This diff
As title

Differential Revision: D86668255

fbshipit-source-id: 6854734ebc349cc07476dcc2df3b70299f5914df