sql_ext: Organize the internal code into a nested module
Summary:
# This stack
Defining a logger config to replace the raw logging scuba we're doing for the XDB telemetry.

My plan is to roll this out, log both ways for a week or two to ensure we don't lose data and then keep only the logger once we know the data looks good.

The main blocker to roll this out now is to be able to capture the logs in unit tests, which would let me use the new logger in the existing tests.

## Why?
Schematized loggers are the best way to log. They provide a lot of features and are more reliable than logging to custom scuba categories.

We'll be able to set custom ODS counters, specify custom sampling rates for scuba and hive, define derived columns, etc.

## This diff

Create a new module called `facebook` and move the functions used by internal build to it.
Doing this before the next diff, because we'll add more of these, along with more fbcode-exclusive imports, so I figured it would be better to tidy everything up in a module.

Differential Revision: D86667661

fbshipit-source-id: 3ce6785e8023e3cc9b8e226778fad5a2ab3317a6