dotgit: write out requirements for submodules
Summary:
Fix Python workingcopy initialization to ensure the requirements are written out for the submodule backing repos. Previously nothing did this for dotgit mode.

This only comes in to play when Sapling has to perform the checkout operation (as opposed to deferring to Git). This happens when there are conflicts during rebase, for example.

Reviewed By: quark-zju

Differential Revision: D86717689

fbshipit-source-id: c4a3326ea45b5358ebc8af7d5f189ac5df7d5096