virtual-repo: limit year in generated commit date
Summary:
When I tested virtual repo with factor bits = 30, the Python side errored out
like:

  ValueError: year 1038681 is out of range

So let's cap the maximum date in generated commits.

Note the `datetime.MAXYEAR` is backed by a C macro constant in CPython
implementation. So setting it to a different value won't change the real limit.

Reviewed By: muirdm

Differential Revision: D86802517

fbshipit-source-id: 7810320032d11ffb7ab53192e5c3a5ded5a31c56