update remote_loader to run fetch in separate thread, have option to background
Summary:
Context:

When hack was running, it would notice random lag spikes of 300-500ms. We added some additional telemetry to try and figure out where it was coming from and identified it to be the network request made when checking for the Use Case config. Every 5 minutes the cache would get expired and it'd need to make a network request again.

This diff:
Moves the network request to a separate thread, and adds a new flag background_fetch.
If this flag is true and there is a cached config, return that immediately and run the fetch in the background.

Reviewed By: muirdm

Differential Revision: D86345036

fbshipit-source-id: deeb4a995955b2d577e5f1951b795c2cb2ee3d2e