sql_ext: Test new logger in telemetry unit tests
Summary:
# This stack
Defining a logger config to replace the raw logging scuba we're doing for the XDB telemetry.

My plan is to roll this out, log both ways for a week or two to ensure we don't lose data and then keep only the logger once we know the data looks good.

The main blocker to roll this out now is to be able to capture the logs in unit tests, which would let me use the new logger in the existing tests.

## Why?
Schematized loggers are the best way to log. They provide a lot of features and are more reliable than logging to custom scuba categories.

We'll be able to set custom ODS counters, specify custom sampling rates for scuba and hive, define derived columns, etc.

## This diff
Use the primitives in D86531442 to capture the schematized logger samples and run the same assertions we run for raw Scuba logging in every unit test.

Differential Revision: D86532523

fbshipit-source-id: 811c716aa92c30f3f8a26f81192522f9ba21edec