commit_graph: add CommitGraphOps to allow generic methods over EdgeType
Summary:
In this stack we are converting the commit graph algorithms to be generic over three possible graph edge types: the regular commit graph using parents, a cut down commit graph using the first parent only, and an expanded commit graph including subtree sources.

In this diff we introduce `CommitGraphOps`, which is a generic over `EdgeType`, and move all methods to this type, making them generic in the process.

The `CommitGraph` now contains a `CommitGraphOps<Parents>`, which behaves the same as before, and derefs to this so that existing code is unchanged.  Additional accessor methods are provided to gain access to `CommitGraphOps` variants for the other two graph types.

Reviewed By: YousefSalama

Differential Revision: D86551370

fbshipit-source-id: 024eed2caaa04d3b6e9b77220828be8c61861005