commit_graph: make ChangesetNode fields private and force use of accessors
Summary:
In this stack we are converting the commit graph algorithms to be generic over three possible graph edge types: the regular commit graph using parents, a cut down commit graph using the first parent only, and an expanded commit graph including subtree sources.

In this diff we make the fields of ChangesetNode private, and force callers to specify which edge type they are interested in when using the appropriate accessors.

Reviewed By: andreacampi

Differential Revision: D86551366

fbshipit-source-id: 5acd95ac1af1f07fb8bdea2e95d5e327f56bb504